TINY-7829: moved all default commands to a separate module
